Start fleshing out the link-to-link network protocol bits.
This is really straightforward stuff; it's supposed to be stupid simple.
Next need a wrapper that abstracts the main application logic and just
gives us a really simple way to do request/response pairs -- in a
magical mysterious place we can build a routing layer on top of that
point-to-point system.