Start sandboxing some stuff.
Got the really basic primitives in; want to muck around a bit with some
stuff before trying to come up with abstractions. Since we're trying to
replace init/etc with a bunch of stuff, spawning processes is pretty
much top of the priority list.

I should make a fuckin' priority list holy shit.