Abstract out the pipes in the sandbox bits.
Decided on NEM_chan_t as the virtual interface for bidirectional message
channels -- it's an interface that it's high-level enough to not care
too much. And it kind of makes sense that you can use the static
dispatch where wanted and use the dynamic dispatch where needed.

Some of the bits need to be abstracted out a bit more, but this works
as-is probably.