Start implementing NEM_unixchan_t.
This does the exact same thing as NEM_pipechan_t -- gives the caller a
pair of linked fds -- but instead of being pipes, they're a (mostly)
anonymous unix domain socket.

This abuses SCM_CREDS to validate that the incoming connection is
originating from the local process. It's a massive cluster right now and
doesn't have things like timeouts -- so it's possible for another
process to come in and start tossing weird data in. Also pretty much
everything except the initial handshake is stubbed out and tests need to
be written and not even NEM_pipechan_t really has tests right now eugh.