Massively refactor (untested) the old channel code.
The previous channel code didn't bother to share any code between
unix/pipe channels which is dumb since they're just file descriptors and
have roughly the same semantics. This breaks the responsibilities into
three separate structs -

 1. NEM_stream_t provides an abstraction around "read N bytes" and
    "write N bytes". Dynamic dispatch.
 2. NEM_fd_t is an implementation of NEM_stream_t that works on plain
    old file descriptors. There will eventually be a TLS one too.
 3. NEM_chan_t adapts a NEM_stream_t to send/recv NEM_msg_t's.

Additionally, some work (not enough) has been put into NEM_app_t -- the
idea is an application framework but honestly I haven't looked at that
code in a couple of days and it's probably stupid and already bitrotted.

And as per tradition there are no tests because this is a huge commit
already. But it compiles! hursehrser.