Start fleshing out some initial command ids.
These are all centralized in a single file for now -- at some point
the proto version might need to be bumped or something to make this
more extensible. For the Go implementation of the service framework
there's a total duplication of all of this, which is kind of a mess.