Rig up NEM_rootd_txnmgr_t.
This is a transaction manager that sits on top of a NEM_stream_t. It
wraps a NEM_chan_t and handles the seq_id matching so that calling code
can just send a message and set an on-response handler. It doesn't
support continued messages or anything yet, but it's passing the initial
tests which is INSANE.

It doesn't do dispatching of incoming requests though -- since txnmgr is
per channel and the dispatcher is probably the same for groups of
channels it didn't make sense to put the logic there.