NEM_rootd_txnmgr_t now takes an unowned NEM_chan_t.
This is to support NEM_child_t which manages it's own NEM_chan_t. It
kind of makes sense that the transaction manager is just a hanger-off of
a channel, and whatever actually owns the channel can use
NEM_chan_on_close to make sure the transaction manager gets freed.