Add NEM_msghdr_pack to marshal msghdrs. And add some tests.
The current allocation strategy is fucking batshit insane because it
requires good-for-nothing pointer fixups. This whole pile of shit is
rigged so that it just allocates a single slab and crashes if the slab
exhausts. This needs to be changed to use a rope-style allocation
strategy with a reasonable starting allocation or something because
having the network message decoding code have panics in it seems like it
might be a goddamn terrible idea.