NEM_msghdr_new: empty strings now become NULL.
This matches the Go implementation which doesn't make a distinction
between NULL and the empty string. It also removes a confusing duality
from the intent of present-but-empty headers. It's possible that, from
the C side, it makes more sense to marshal them as "" instead? But it
makes the unmarshalling code a bit more complex due to libbson requiring
object iteration for value retrieval.

Also make the filenames in nem-rootd/src conform.