nem-rootd: skeleton out svc-host implementation.
Need to rework svcmgr to be a bit more like the Go implementation --
currently there's no way to chain multiple services together without
modifying the svc definition. This is kind of ugly; it makes more sense
to remove the 'next' functionality from the svcmgr and provide a
separate abstraction for combining them (since it's likely that we'll
use that functionality to add allowed services to a connection
post-authorization).