nem-rootd: hook in sqlite3.
The local image registry needs to persist structured data. I want to
keep versions of images and stuff (and have symbolic names for a set of
versions) and it makes more sense to do that within sqlite3 than flat
files or something equally inane. The runtime state can be collected
from the kernel -- this is more for the on-disk image metadata.