c-rootd: imgset: reorganize how the image sets are laid out.
The previous one had versions strictly owned by images. This kind of
didn't make sense because "versions" are just a SHA256-identified file
and "images" are just a symbolic tie around a set of "versions". So a
version could belong to multiple "images".

This changes it so they're stored as a many-to-many relationship, so
that image files can be shared between "images".