libnem: Initial JSON unmarshalling, untested.
Also starting to put together a more robust testing framework for the
marshalling bits. The goal is to have a set of types that flex all of
the functionality and to pull them in with a pile of macros (since the
marshalling bits have a lot of fuckin' type combinations). This should
be significantly more robust than just using a handful of types like the
old tests did.