nem-rootd: start hooking up router initialization.
Not entirely convinced this is the best way to handle it; daemon/setcfg
might be a better approach. If the router is going to be running under
capsicum it can't bind ports -- so the root needs to handle that. Which
is probably fine since the root can bind ports and wedge 'em open with
an extra ref.

Probably going to need to rewrite reams of this.

In any case, currently a bug in the IPC somewhere wherein the router
doesn't restart correctly (seemingly random).