libnem: NEM_chan_send now takes an optional completion callback.
This also changes the scheduling on NEM_chan_send to be FIFO rather than
LIFO -- while LIFO has some really nice properties, an unstable ordering
is a recipe for disaster; especially when we're using flags to indicate
message stream start/end. With a LIFO queue we'd need to add another
header and lever a bunch of complexity onto the caller to reorder
messages. Not really convinced that's better than just eating the head
of line blocking semantics.