libnem: txnmgr: unfuck multiple-request transactions a bit.
The previous design was a bit crazy; it required the service handler for
a multiple-request transaction to explicitly set NEM_txnin_t.thunk to
some value to process additional requests. This is ... kind of stupid.
It makes a lot more sense to instead just repeatedly pass new messages
to the same handler (given they can stuff data in the NEM_txnin_t
directly).