libnem: txnmgr: some timeout tests, fixing bugs.
There were a whole bunch of data ownership issues in the original design
(re: what is the lifetime of a NEM_txn_t*?). This changes the code to
ensure that any NEM_txn_t* the caller has access to is valid until a
final req/reply is sent, the transaction is explicitly cancelled by the
caller, or the manager is free'd.