libnem: add NEM_app_t tests.
This juggles around the in-app kq initialization to occur when the app
is initialized (rather than after component setups have been run). If
it's not initialized when the setups are run, the setups can't actually
do anything.

As a note, nem-rootd and friends are currently horribly broken. Need to
go back and port them over to the new (improved) systems.