nem-hostd: figuring out the configuration file bits.
Doing this slightly differently this time; want the hostd to be
controlled by a single configuration file (that lives in a git repo or
elsewhere). When the configuration file gets updated, it's reloaded by
the hostd and the changes are applied to the working state.

This doesn't really work with automatic orchestration, but that's
looking like a totally unnecessary component.

And now I'm looking at integrating libyaml because I'm not sure toml is
really a good fit for some of the structures I want to have in the
config file :(