libnem: add NEM_file_t for reading files.
This is for read-only files; for writing files probably better to just
open+write+close directly (and for structured files... sqlite3). This
wraps up the mmap semantics and error handling. At some point it
probably makes sense to hook it into kq and let the file be watched for
updates.