nem-rootd: remove the state component.
This bit is trying to do a lot that isn't needed -- the goal was to
centralize all the running state so it could be marshalled over a unix
domain socket to a new rootd instance so that the new one could continue
running. This is a noble goal (seamless updates) but can add it later.
The only thing we really need to keep around are the jail fds, the rest
of the runtime state should be recoverable from system scanning.