nem-rootd: pulling bits together in c-mounts.
Really not entirely happy with how these components are integrating;
feels kind of soupy. There's a hell of a lot of missing subtlety from
each of the individual components that's only starting to leak out now
that I need to actually tie them together.

Throwing a WIP commit in here because I need to sort out the semver bits
and that'll probably be a reasonably sized commit.